requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
gunicorn>=21.2
whitenoise>=6.6
python-decouple>=3.8
django-cors-headers>=4.3
Pillow>=10.0